+++ Nos/Views/Note/NoteCardHeader.swift
@@ -1,24 +1,35 @@
+import Combine
 import SwiftUI
 
 struct NoteCardHeader: View {
     
-    @ObservedObject var note: Event
-    @ObservedObject var author: Author
+    let authorSafeName: String
+    let authorProfilePhotoURL: URL?
+    let noteExpirationDate: Date?
+    let noteCreatedDate: Date?
+    
+    @State private var currentTime = Date.now
+    @State private var expirationDateDistanceString: String?
+    @State private var createdDateDistanceString: String?
+    
+    /// A timer used to cause a periodic refresh of the date strings.
+    @State private var timer: Timer.TimerPublisher?
+    @State private var timerCancellable: AnyCancellable?
     
     var body: some View {
         HStack(alignment: .center, spacing: 8) {
-            AuthorLabel(author: author, note: note)
-            if let expirationTime = note.expirationDate?.distanceString() {
+            AuthorLabel(name: authorSafeName, profilePhotoURL: authorProfilePhotoURL)
+            if let expirationDateDistanceString {
                 Image.disappearingMessages
                     .resizable()
                     .foregroundColor(.secondaryTxt)
                     .frame(width: 25, height: 25)
-                Text(expirationTime)
+                Text(expirationDateDistanceString)
                     .lineLimit(1)
                     .font(.clarity(.medium))
                     .foregroundColor(.secondaryTxt)
-            } else if let elapsedTime = note.createdAt?.distanceString() {
-                Text(elapsedTime)
+            } else if let createdDateDistanceString {
+                Text(createdDateDistanceString)
                     .lineLimit(1)
                     .font(.clarity(.medium))
                     .foregroundColor(.secondaryTxt)
@@ -27,15 +38,57 @@ struct NoteCardHeader: View {
             Spacer()
         }
         .padding(.leading, 10)
+        .onAppear {
+            update()
+            startTimer()
+        }
+        .onDisappear {
+            stopTimer()
+        }
+    }
+    
+    private func update() {
+        createdDateDistanceString = noteCreatedDate?.distanceString()
+        if let noteExpirationDate {
+            expirationDateDistanceString = currentTime.distanceString(noteExpirationDate)
+        }
+    }
+    
+    private func startTimer() {
+        timer = Timer.publish(every: 60, on: .main, in: .common)
+        timerCancellable = timer?
+            .autoconnect()
+            .sink { newTime in
+                currentTime = newTime
+                update()
+            }
+    }
+    
+    private func stopTimer() {
+        timerCancellable?.cancel()
+        timerCancellable = nil
+        timer = nil
     }
 }
 
 struct NoteCardHeader_Previews: PreviewProvider {
     static var previewData = PreviewData()
+    
     static var previews: some View {
-        NoteCardHeader(note: previewData.imageNote, author: previewData.previewAuthor)
-            .inject(previewData: previewData)
-        NoteCardHeader(note: previewData.expiringNote, author: previewData.previewAuthor)
-            .inject(previewData: previewData)
+        NoteCardHeader(
+            authorSafeName: previewData.previewAuthor.safeName,
+            authorProfilePhotoURL: previewData.previewAuthor.profilePhotoURL,
+            noteExpirationDate: previewData.imageNote.expirationDate,
+            noteCreatedDate: previewData.imageNote.createdAt
+        )
+        .inject(previewData: previewData)
+        
+        NoteCardHeader(
+            authorSafeName: previewData.previewAuthor.safeName,
+            authorProfilePhotoURL: previewData.previewAuthor.profilePhotoURL,
+            noteExpirationDate: previewData.expiringNote.expirationDate,
+            noteCreatedDate: previewData.expiringNote.createdAt
+        )
+        .inject(previewData: previewData)
     }
 }