+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed: crash in Event.trackDelete(on:context:).
 
 ### Internal Changes
+- Fixed: Contact Support event fires too often.
+- Performance improvements for RepliesLabel, AuthorLabel, NoteCardHeader, Date+Elapsed
 
 ## [1.2.1] - 2025-02-19Z